Fix Generation of Sub-Keys in Dictionaries
The tool was previously non-deterministic in code generation for
different orders of property lists passed as parameters. This
generated some side effects where blocks of code wouldn’t be generated
at all depending on the order of property lists that were passed to
the tool **and** the difference in keys between these property lists.

This commit introduces a smarter behavior for property lists that
contain dictionaries, and these dictionaries contain different keys,
also called sub-keys in this context. Before code generation already
known keys that aren’t present in the currently processed property
list are merged with its contents, so that further code generation
methods can determine the optionality of these keys. These keys were
previously not part of the generated code.