Add Package Structure for Better Development Integration
It is possible to generate a local Xcode project for a local
development environment now: `swift package generate-xcodeproj`. The
project won’t be commited and you can debug with custom parameters set
in the scheme’s argument.

It is no longer necessary to compile the tool before usage. It can
be used as a regular script.